made creategame a component
